requirements.txt
python-dateutil
Pillow
mysql.connector